generate-report-lambda/report_code_old/requirements.txt
greenlet==3.0.3
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
scipy==1.13.1
docxtpl